Class for merging moved to rnftools.rnfformat
